Moved handle_click_event from script_task.rs to document.rs
